+++ src/base/vgeu_gltf.hpp
@@ -16,13 +16,11 @@ https://github.com/SaschaWillems/Vulkan/blob/master/base/VulkanglTFModel.h
 #include "tiny_gltf.h"
 #define GLM_FORCE_RADIANS
 #define GLM_FORCE_DEPTH_ZERO_TO_ONE
+#include <Vulkan-Hpp/vulkan/vulkan.hpp>
+#include <Vulkan-Hpp/vulkan/vulkan_raii.hpp>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-
-//
-#include <Vulkan-Hpp/vulkan/vulkan.hpp>
-#include <Vulkan-Hpp/vulkan/vulkan_raii.hpp>
 
 // std
 #include <limits>
@@ -66,42 +64,6 @@ struct FlagTraits<RenderFlagBits> {
 namespace glTF {
 
 struct Node;
-
-// modified existing structure to fit in RAII paradigm.
-struct Texture {
-  std::unique_ptr<vgeu::VgeuImage> vgeuImage;
-  vk::ImageLayout imageLayout{};
-  uint32_t width = 0;
-  uint32_t height = 0;
-  uint32_t mipLevels = 0;
-  uint32_t layerCount = 0;
-  vk::DescriptorImageInfo descriptorInfo{};
-  vk::raii::Sampler sampler = nullptr;
-
-  // fromglTFImage
-  Texture(tinygltf::Image& gltfimage, std::string path,
-          const vk::raii::Device& device, VmaAllocator allocator,
-          const vk::raii::Queue& transferQueue,
-          const vk::raii::CommandPool& commandPool);
-  // empty texture
-  Texture(const vk::raii::Device& device, VmaAllocator allocator,
-          const vk::raii::Queue& transferQueue,
-          const vk::raii::CommandPool& commandPool);
-
-  void fromglTFImage(tinygltf::Image& gltfimage, std::string path,
-                     const vk::raii::Device& device, VmaAllocator allocator,
-                     const vk::raii::Queue& transferQueue,
-                     const vk::raii::CommandPool& commandPool);
-  void generateMipmaps(const vk::raii::CommandBuffer& cmdBuffer);
-  void createEmptyTexture(const vk::raii::Device& device,
-                          VmaAllocator allocator,
-                          const vk::raii::Queue& transferQueue,
-                          const vk::raii::CommandPool& commandPool);
-  // NOTE: use mipLevels
-  void createSampler(const vk::raii::Device& device);
-  // NOTE: used at the end of fromglTFImage()
-  void updateDescriptorInfo();
-};
 
 struct Material {
   enum class AlphaMode { kALPHAMODE_OPAQUE, kALPHAMODE_MASK, kALPHAMODE_BLEND };
