+++ src/base/vgeu_texture.cpp
@@ -1,0 +1,1 @@
+#include "vgeu_texture.cpp"
